requirements.txt
numpy<2
scipy==1.10.* 
networkx==2.8.8 
shapely==2.0.* 
pyproj==3.6.* 
rtree==1.2.* 
fiona==1.9.* 
geopandas==0.14.* 
osmnx==1.9.4 
earthengine-api>=0.1.398
fastapi
uvicorn[standard]
websockets
pydantic
folium
google-auth>=2.30.0
python-dotenv
groq
markdown
markdownify
aiofiles
python-multipart